Multiple changes, mostly to maintain compatibility with Octave.
The function readfis was modified to workaround the change to
strsplit beginning in Octave 3.8.0. Without the modification,
readfis will not work with Octave versions >= 3.8.0. The new
version of readfis works with all versions of Octave >= 3.2.4
by first checking for the version number of Octave and then
selecting either ostrsplit (for Octave >= 3.8.0) or strsplit
(for Octave < 3.8.0).

The files writefis.m and evalmf.m were edited to maintain
compatibility with future versions of Octave. Two occurrences of
the continuation "..." within double quoted strings in writefis.m
were changed to "\". One occurrence of "..." in evalmf.m was
removed by writing the instruction on a single line.

Five private functions renamed on 2014-03-06 were restored to
their original names, because renaming them breaks gustafson_kessel.m.

The file CITATION was added to include references for two papers
written about the fuzzy-logic-toolkit.

All copyright notices were updated.